# cmd/web/docs/swagger.yaml
definitions:
  handler.TaskRequest:
    properties:
      expire_days:
        type: integer
      priority:
        maximum: 4
        minimum: 0
        type: integer
      text:
        type: string
      title:
        minLength: 1
        type: string
    required:
    - title
    type: object
  handler.TaskResponse:
    properties:
      expires:
        type: string
      id:
        minimum: 1
        type: integer
      priority:
        maximum: 4
        minimum: 0
        type: integer
      text:
        type: string
      title:
        minLength: 1
        type: string
    required:
    - id
    - title
    type: object
  handler.UserRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      role:
        enum:
        - user
        - admin
        type: string
      username:
        type: string
    type: object
  handler.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        minimum: 1
        type: integer
      role:
        enum:
        - user
        - admin
        type: string
      username:
        type: string
    required:
    - id
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      - multipart/form-data
      description: login to get token for accessing other endpoints
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        minLength: 8
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token for accessing endpoints
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: login
      tags:
      - auth
  /tasks:
    get:
      consumes:
      - application/json
      description: list tasks from database using offset and limit
      parameters:
      - default: 5
        description: Limit for tasks listing
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for tasks listing
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.TaskResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: list tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: add task into database
      parameters:
      - description: task schema for creating a new user
        in: body
        name: tasks
        required: true
        schema:
          $ref: '#/definitions/handler.TaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/handler.TaskResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: add task
      tags:
      - tasks
  /tasks/:id:
    delete:
      consumes:
      - application/json
      description: delete task from database
      parameters:
      - description: id for deleting a task
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns a confirmation message
          schema:
            additionalProperties:
              type: string
            type: object
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: delete task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: get task from database by id
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TaskResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: get task
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: patch task from database
      parameters:
      - description: id for patching a task
        in: path
        name: id
        required: true
        type: integer
      - description: task schema for patching a task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/handler.TaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns a confirmation message
          schema:
            additionalProperties:
              type: string
            type: object
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: patch task
      tags:
      - tasks
  /users:
    get:
      consumes:
      - application/json
      description: get users from database by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.UserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: get users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: create a new user
      parameters:
      - description: Credentials to signup
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handler.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token for accessing endpoints
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: signup
      tags:
      - auth
  /users/:id:
    delete:
      consumes:
      - application/json
      description: delete user from database by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: get user from database by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: get user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: patch user from database by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: user schema for patching a user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: patch user
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"